+++ CHANGELOG.md
@@ -13,6 +13,10 @@ All notable changes to this project will be documented in this file.
 * When entering edit mode, Sabaki will now focus the comment box automatically
 * Send custom initial GTP commands to engines after sending basic meta commands
 
+**Fixed**
+
+* Correctly send handicap stones to GTP engines
+
 ## [Sabaki v0.34.1][v0.34.1] (2018-06-03)
 
 **Changed**